more duplicate code removal
