+++ client/src/components/Auth/BusinessSignup.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import { UserType } from '../../utils/types'
 import { useNavigate } from 'react-router-dom'
 import { containerStyle, formFieldOnChange, hasError } from '../../utils/utils'
-import AuthTitle from '../Shared/AuthTitle'
+import AuthTitle from './AuthTitle'
 import Grid from '@mui/material/Grid'
 import Button from '@mui/material/Button'
 import Link from '@mui/material/Link'
