+++ client/src/components/Home/index.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react'
-import BusinessHome from './BusinessHome'
 import { UserType } from '../../utils/types'
-import InfluencerHome from './InfluencerHome'
 import { isAuthenticated } from '../../utils/utils'
 import useAuth from '../../hooks/AuthHook'
+import MarketingTaskList from './MarketingTaskList'
+import HomePageGrid from './HomePageGrid'
+import ApplicationList from './ApplicationList'
 
 export default function Home (): JSX.Element {
   const auth = useAuth()
@@ -11,7 +12,10 @@ export default function Home (): JSX.Element {
   return (
     <React.Fragment>
       {isAuthenticated(auth.userType)
-        ? (auth.userType === UserType.BUSINESS ? <BusinessHome/> : <InfluencerHome/>)
+        ? (auth.userType === UserType.BUSINESS
+            ? <HomePageGrid listComp={<MarketingTaskList/>}/>
+            : <HomePageGrid listComp={<ApplicationList/>}/>
+          )
         : <div>Hello World</div>
       }
     </React.Fragment>
